// src/resources/dto/GetSettingsTargetResDto.ts
/** Copyright 2022 NEC Corporation
Released under the MIT license.
https://opensource.org/licenses/mit-license.php
*/
/**
 *
 *
 *
 * $Date$
 * $Revision$
 * $Author$
 *
 * TEMPLATE VERSION :  76463
 */
export class CodeObject {
    _value: number;

    _ver: number;
}

export default class GetSettingsTargetResDto {
    pxrId: string;

    dataOperations: CodeObject[];
}